client: gui: RoadPainter: optimize road rendering
The previous implementation first generated a list of applicable
bordering directions (via streams and extra predicate object calls),
and then again iterates this list to collect the corner points.

This can be easily done in one pass, as the second list is a direct
mapping out of the first one (IOW: the same decision criteria).

And the data set is so small that the overhead of streams and
MethodRef calls really outweights the potetial gain from parallelization.
OTOH, we could think about rendering several tiles in parallel.