common: model: Player: optimize visibiliy map
The old approach required lots locking (for each individual test)
and lots of memory (array of arrays of boolean). Replacing this by
an memory efficient BitSet, which is allocated and computed on-demand
and just dropped for invalidation.

Additionally only hold a lock when the visibility map needs to be
recomputed. The lock is held on the Player object itself, so we dont
need an additional lock object anymore (note: the visibility map may
be null). For now, it's the only case for that (other parts lock
referred objects), so there's no risk of penalty by this global lock.