common: model: UnitType: cache heavily used ability flags
Checking for abilities is a rather expensive operation, and it's
used quite a lot (eg. on each unit paint cycle). Some of UnitType's
abilities are pretty constant (basicly, just change when the UnitType
is (re)loaded from xml). So, we better cache them in local fields
to have the read access as fast as possible.