common: model: FreeColObject: optimize-away static applyModifiers()
The applyModifiers() class methods are just wrappers for the corresponding
FreeColContainer.applyModifiers() methods, so we can call them directly
and so clean up FCO's interace a little bit.

Also, in some cases we just apply it to the modifiers from another
FCO - its cleaner and quicker to call its applyModifiers() method directly.